requirements.txt
fastapi~=0.128.0
uvicorn[standard]==0.30.*
SQLAlchemy==2.0.*
asyncpg==0.29.*
alembic==1.13.*
python-dotenv==1.0.*
pydantic-settings==2.4.*
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.2.*
pydantic~=2.12.5